Small changes in preparation for battle system.
- Remove obsolete Discord bot token used for early testing.

- Add BasicSpeed() to Entity class.

- Change all doubles to decimals.

- Change JestersCap to only be usable by Fool character class.

+++ Equipment.cs
@@ -111,7 +111,7 @@ namespace DURPSBot
         {
             return prefix + " " + Name + " " + suffix;
         }
-        public double TotaWeight()
+        public decimal TotaWeight()
         {
             return Weight + (Weight * Prefix.WeightMultiplier - Weight) + (Weight * Suffix.WeightMultiplier - Weight);
         }
